+++ tests/blueprint_typestate_option.rs
@@ -49,6 +49,14 @@ impl<OPTION,> NamedBuilder<(),OPTION,> {
 // impl_builder / impl_builder_setter
 impl<MANDATORY,> NamedBuilder<MANDATORY,(),> {
 
+    pub fn option_optional<OPTION: ::core::convert::Into<PathBuf>>(self, option: ::core::option::Option<OPTION>) -> NamedBuilder<MANDATORY,Option<PathBuf>,> {
+        NamedBuilder {
+            mandatory: self.mandatory,
+            option: option.map(::core::convert::Into::into),
+            __typestate_markers: ::core::default::Default::default(),
+        }
+    }
+
     pub fn option<OPTION: ::core::convert::Into<PathBuf>>(self, option: OPTION) -> NamedBuilder<MANDATORY,Option<PathBuf>,> {
         NamedBuilder {
             mandatory: self.mandatory,
@@ -117,6 +125,14 @@ impl TupleBuilder<(),()> {
 // impl_builder / impl_builder_setter
 impl<V0,> TupleBuilder<V0,(),> {
 
+    pub fn set1_optional<V1: ::core::convert::Into<PathBuf>>(self, v1: ::core::option::Option<V1>) -> TupleBuilder<V0,Option<PathBuf>,> {
+        TupleBuilder(
+            self.0,
+            v1.map(::core::convert::Into::into),
+            ::core::default::Default::default(),
+        )
+    }
+
     pub fn set1<V1: ::core::convert::Into<PathBuf>>(self, v1: V1) -> TupleBuilder<V0,Option<PathBuf>,> {
         TupleBuilder(
             self.0,
@@ -134,6 +150,9 @@ impl<V0,> TupleBuilder<V0,(),> {
     }
 }
 impl TupleBuilder<PathBuf,()> {
+    pub fn optional<V1: ::core::convert::Into<PathBuf>>(self, v1: ::core::option::Option<V1>) -> TupleBuilder<PathBuf,Option<PathBuf>,> {
+        self.set1_optional(v1)
+    }
     pub fn set<V1: ::core::convert::Into<PathBuf>>(self, v1: V1) -> TupleBuilder<PathBuf,Option<PathBuf>,> {
         self.set1(v1)
     }
@@ -179,6 +198,21 @@ fn named_builder_build_full() {
 }
 
 #[test]
+fn named_builder_build_full_optional() {
+    let built = Named::builder()
+        .option_optional(Some("/tmp/builder_build_full/option"))
+        .mandatory("/tmp/builder_build_full/mandatory")
+        .build();
+    assert_eq!(
+        Named {
+            mandatory: PathBuf::from("/tmp/builder_build_full/mandatory"),
+            option: Some(PathBuf::from("/tmp/builder_build_full/option")),
+        },
+        built,
+    );
+}
+
+#[test]
 fn named_builder_build_partial_implicit() {
     let built = Named::builder()
         .mandatory("/tmp/builder_build_partial_implicit/mandatory")
@@ -206,11 +240,27 @@ fn named_builder_build_partial_explicit() {
         built,
     );
 }
+
 #[test]
 fn named_builder_into_full() {
     let built = Named::builder()
         .mandatory("/tmp/builder_into_full/mandatory")
         .option("/tmp/builder_into_full/option")
+        .into();
+    assert_eq!(
+        Named {
+            mandatory: PathBuf::from("/tmp/builder_into_full/mandatory"),
+            option: Some(PathBuf::from("/tmp/builder_into_full/option")),
+        },
+        built,
+    );
+}
+
+#[test]
+fn named_builder_into_full_optional() {
+    let built = Named::builder()
+        .mandatory("/tmp/builder_into_full/mandatory")
+        .option_optional(Some("/tmp/builder_into_full/option"))
         .into();
     assert_eq!(
         Named {
@@ -254,6 +304,21 @@ fn named_builder_into_partial_explicit() {
 fn tuple_builder_build_unordered_full() {
     let built = Tuple::builder()
         .set1("/tmp/builder_build_full/option")
+        .set0("/tmp/builder_build_full/mandatory")
+        .build();
+    assert_eq!(
+        Tuple(
+            PathBuf::from("/tmp/builder_build_full/mandatory"),
+            Some(PathBuf::from("/tmp/builder_build_full/option")),
+        ),
+        built,
+    );
+}
+
+#[test]
+fn tuple_builder_build_unordered_full_optional() {
+    let built = Tuple::builder()
+        .set1_optional(Some("/tmp/builder_build_full/option"))
         .set0("/tmp/builder_build_full/mandatory")
         .build();
     assert_eq!(
@@ -310,6 +375,21 @@ fn tuple_builder_build_ordered_full() {
 }
 
 #[test]
+fn tuple_builder_build_ordered_full_optional() {
+    let built = Tuple::builder()
+        .set("/tmp/builder_build_full/mandatory")
+        .optional(Some("/tmp/builder_build_full/option"))
+        .build();
+    assert_eq!(
+        Tuple(
+            PathBuf::from("/tmp/builder_build_full/mandatory"),
+            Some(PathBuf::from("/tmp/builder_build_full/option")),
+        ),
+        built,
+    );
+}
+
+#[test]
 fn tuple_builder_build_ordered_partial_implicit() {
     let built = Tuple::builder()
         .set("/tmp/builder_build_partial_implicit/mandatory")
@@ -343,6 +423,21 @@ fn tuple_builder_into_full() {
     let built = Tuple::builder()
         .set0("/tmp/builder_into_full/mandatory")
         .set1("/tmp/builder_into_full/option")
+        .into();
+    assert_eq!(
+        Tuple(
+            PathBuf::from("/tmp/builder_into_full/mandatory"),
+            Some(PathBuf::from("/tmp/builder_into_full/option")),
+        ),
+        built,
+    );
+}
+
+#[test]
+fn tuple_builder_into_full_optional() {
+    let built = Tuple::builder()
+        .set0("/tmp/builder_into_full/mandatory")
+        .set1_optional(Some("/tmp/builder_into_full/option"))
         .into();
     assert_eq!(
         Tuple(