+++ tests/macro_result_option.rs
@@ -43,6 +43,21 @@ fn named_builder_build_full() {
 }
 
 #[test]
+fn named_builder_build_full_optional() {
+    let built = Named::builder()
+        .option_optional(Some("/tmp/builder_build_full/option"))
+        .mandatory("/tmp/builder_build_full/mandatory")
+        .build();
+    assert_eq!(
+        Ok(Named {
+            mandatory: PathBuf::from("/tmp/builder_build_full/mandatory"),
+            option: Some(PathBuf::from("/tmp/builder_build_full/option")),
+        }),
+        built,
+    );
+}
+
+#[test]
 fn named_builder_build_partial_implicit() {
     let built = Named::builder()
         .mandatory("/tmp/builder_build_partial_implicit/mandatory")
@@ -70,11 +85,27 @@ fn named_builder_build_partial_explicit() {
         built,
     );
 }
+
 #[test]
 fn named_builder_into_full() {
     let built = Named::builder()
         .mandatory("/tmp/builder_into_full/mandatory")
         .option("/tmp/builder_into_full/option")
+        .try_into();
+    assert_eq!(
+        Ok(Named {
+            mandatory: PathBuf::from("/tmp/builder_into_full/mandatory"),
+            option: Some(PathBuf::from("/tmp/builder_into_full/option")),
+        }),
+        built,
+    );
+}
+
+#[test]
+fn named_builder_into_full_optional() {
+    let built = Named::builder()
+        .mandatory("/tmp/builder_into_full/mandatory")
+        .option_optional(Some("/tmp/builder_into_full/option"))
         .try_into();
     assert_eq!(
         Ok(Named {
@@ -130,6 +161,21 @@ fn tuple_builder_build_full() {
 }
 
 #[test]
+fn tuple_builder_build_full_optional() {
+    let built = Tuple::builder()
+        .set1_optional(Some("/tmp/builder_build_full/option"))
+        .set0("/tmp/builder_build_full/mandatory")
+        .build();
+    assert_eq!(
+        Ok(Tuple(
+            PathBuf::from("/tmp/builder_build_full/mandatory"),
+            Some(PathBuf::from("/tmp/builder_build_full/option")),
+        )),
+        built,
+    );
+}
+
+#[test]
 fn tuple_builder_build_partial_implicit() {
     let built = Tuple::builder()
         .set0("/tmp/builder_build_partial_implicit/mandatory")
@@ -163,6 +209,21 @@ fn tuple_builder_into_full() {
     let built = Tuple::builder()
         .set0("/tmp/builder_into_full/mandatory")
         .set1("/tmp/builder_into_full/option")
+        .try_into();
+    assert_eq!(
+        Ok(Tuple(
+            PathBuf::from("/tmp/builder_into_full/mandatory"),
+            Some(PathBuf::from("/tmp/builder_into_full/option")),
+        )),
+        built,
+    );
+}
+
+#[test]
+fn tuple_builder_into_full_optional() {
+    let built = Tuple::builder()
+        .set0("/tmp/builder_into_full/mandatory")
+        .set1_optional(Some("/tmp/builder_into_full/option"))
         .try_into();
     assert_eq!(
         Ok(Tuple(